+++ fl_outgoing_mail_server_user/models/ir_mail_server.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from odoo import api, fields, models
-
-
-class IrMailServer(models.Model):
-    _inherit = 'ir.mail_server'
-
-    user_id = fields.Many2one('res.users', string='User')