+++ klo_contract_line_menu/__manifest__.py
@@ -1,0 +1,17 @@
+# Copyright 2024 Manuel Calomarde Gómez - KLO Ingeniería Informática S.L.L.
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+{
+    "name": "KLO Contract Line Menu",
+    "version": "15.0.1.0.0",
+    "category": "Contract Management",
+    "license": "AGPL-3",
+    "author": "KLO Ingeniería Informática S.L.L.",
+    "website": "https://www.klo.es",
+    "depends": [
+        "contract",
+    ],
+    "data": [
+        "views/contract_line_views.xml",
+    ],
+    "installable": True,
+}