+++ klo_unit_cost_stock_valuation_layer/__manifest__.py
@@ -1,0 +1,18 @@
+# Copyright 2024 Manuel Calomarde Gómez - KLO Ingeniería Informática S.L.L.
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+{
+    "name": "KLO Unit cost on Stock valuation layer",
+    "version": "15.0.1.0.0",
+    "category": "Warehouse Management",
+    "license": "AGPL-3",
+    "author": "KLO Ingeniería Informática S.L.L.",
+    "website": "https://www.klo.es",
+    "depends": [
+        "stock_account",
+    ],
+    "data": [
+        "data/ir_action_data.xml",
+        "views/stock_valuation_layer_views.xml",
+    ],
+    "installable": True,
+}