+++ sh_night_mode/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (C) Softhealer Technologies.
-
-from . import models