+++ snazzy_theme_common/controllers/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-# Developed by Bizople Solutions Pvt. Ltd.
-# See LICENSE file for full copyright and licensing details
-
-from . import main
-from . import dynamic_config_tool
-from . import portal_address
-from . import store_location