+++ stock_inventory_adjustment/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from . import stock_inventory
+from . import product_product
+from . import stock_inventory_line
+from . import res_users
+from . import xls