+++ prt_report_attachment_preview/doc/changelog.rst
@@ -1,9 +1,24 @@
-14.0.1.0.1
+15.0.1.0.4
 -----------
 
-- Bug fixes
+- Fix: bug fixes
 
-14.0.1.0.0
+15.0.1.0.3
 -----------
 
-- Release for Odoo 14
+- Fix: repeated button actions
+
+15.0.1.0.2
+-----------
+
+- Fix: printing via action
+
+15.0.1.0.1
+-----------
+
+- Fix: product labels printing
+
+15.0.1.0.0
+-----------
+
+- Release for Odoo 15